refactor(zero_trust_list): isolate custom logic into items.go
Move all hand-written logic (fetchItems, suppressItemsDiff, computeItemsHash,
helpers) into items.go so it is clearly separated from the generated resource.go.

resource.go now has two minimal touch points:
  - Read()       calls fetchItems() — one line
  - ModifyPlan() delegates to suppressItemsDiff() — one line

This makes the generated file easy to re-apply on future regenerations without
losing any of our custom behaviour.
